Move GasUpdate to new file
