+++ crates/starknet-devnet-types/src/rpc.rs
@@ -5,6 +5,7 @@ pub mod emitted_event;
 pub mod estimate_message_fee;
 pub mod eth_address;
 pub mod felt;
+pub mod gas_update;
 mod macro_utils;
 pub mod messaging;
 pub mod state;
